\name{pangui} 
\alias{pangui} 
\title{ start a graphical user interface } 
\description{ 
    A function which start a graphical  user interface to write graphical code chunks. 
} 
\usage{ pangui(infile=NULL,quiet=FALSE) } 
\arguments{ 
  \item{infile}{ 
    an infile to be edited with  the graphical interface, default: NULL 
  } 
  \item{quiet}{ 
    should messages been hidden, default: FALSE 
  } 
} 
\details{ 
    Some more details ... 
} 
\examples{ 
\dontrun{ 
  pangui() 
} 
} 
